refactor: Renamed Team component
